updated demo
set changed to regenerate which creates items between the start and end
adding a new item to the timeline now respects the demo durations
removed the guard condition on willReceiveProps to always re-render
